Fix repeats in output when cmd line objects depend on each other.
If an object on the command line depends on another object the output
code wrote the object files and then the dependent files and a
command line object file that is dependent ended up in the object
and dependent lists. A simple merge and unique fixed it.

Also moved the script generation code into a separate function
that can be used in a application container.