Output application format files.
Added support for an RTEMS RAP format application file. The format is:

 <header>
 <LZ77>
  <Application Script>
  <[1..n] ELF Object files>
 </LZ77>

Where the header is a text string of fields delimited by ',' and terminated
with a line feed (\n). It is variable length:

  RTEMS-APP,0000000,01.00.00,LZ77,00000000\n\0

  where:
   RTEMS-APP     : file tag for quick acceptance and rejection
   Length        : the length of the application in bytes including the
                 : header
   Version       : Version of the application format.
   Compress      : The compression format.
   Checksum      : CCITT CRC32 checksum.

Following the header is a nul ('\0') character then an LZ77 container
with the application loader script followed by the ELF object files.

Note, the script format will be documented else where.

Note, the final version may add a 32bit length field before each part
in the compressed container to delimit the size of the file to be
read. This is currently not in this version.