Make layout offsets match image offsets. Add a better strtab search.
The offsets of sections in objects have to be computed in load order
so the offsets work. For example an object could have a sections with
an alignment of 4 and 16 so the 4 may align with the current load
offset and the 16 is padded to align it. This is the way the image is
loaded and so the layout needs to follow these rules.

The strtab search needs to be better. An STL string find was finding
patrial strings and causing errors on the target.