Fix RELA in RAP formats.
Fixed the RELA handling in the merging of the object sections into the
RAP sections. A method in the rap::image class was not being used any
more and that code transferred the RELA setting from the object sections
to the image section. This change adds also transfers the rela field.